ref: Move redis abstraction to own module
